Crate jrsonnet_evaluator

jsonnet interpreter implementation

Re-exports

pub use error::Error;

pub use error::Result;

pub use error::ResultExt;

pub use val::Thunk;

pub use val::Val;

pub use jrsonnet_parser as parser;

pub use error::ErrorKind::*;

Modules

destructure

error

function

gc

manifest

operator

stack

stdlib

trace

typed

val

Macros

Either

bail

runtime_error

tb

Structs

Context

Context keeps information about current lexical code location

ContextBuilder

DummyImportResolver

Dummy resolver, can’t resolve/load any file

EvaluationSettings

Dynamically reconfigurable evaluation settings

EvaluationStateInternals

ExtendBuilder

FileImportResolver

File resolver, can load file from both FS and library paths

IBytes

Interned byte array

IStr

Interned string

InitialUnderscore

ObjFieldFlags

ObjMember

ObjMemberBuilder

ObjValue

ObjValueBuilder

OopObject

Pending

State

Maintains stack trace and import resolution

ValueBuilder

WeakObjValue

Enums

MaybeUnbound

Object fields may, or may not depend on this/super, this enum allows cheaper reuse of object-independent fields for native code Standard jsonnet fields are always unbound

Traits

ContextInitializer

During import, this trait will be called to create initial context for file. It may initialize global variables, stdlib for example.

ImportResolver

Implements file resolution logic for import and importStr

ObjectAssertion

ObjectLike

Unbound

Thunk without bound super/this object inheritance may be overriden multiple times, and will be fixed only on field read

Functions

apply_tla

evaluate

evaluate_apply

evaluate_assert

evaluate_comp

evaluate_field_member

evaluate_field_name

evaluate_member_list_object

evaluate_method

evaluate_named

evaluate_object

evaluate_trivial