Module jrsonnet_evaluator::function

Re-exports

pub use arglike::ArgLike;

pub use arglike::ArgsLike;

pub use arglike::TlaArg;

Modules

arglike

builtin

native

parse

Structs

CallLocation

Function callsite location. Either from other jsonnet code, specified by expression location, or from native (without location).

FuncDesc

Represents Jsonnet function defined in code.

Enums

FuncVal

Represents a Jsonnet function value, including plain functions and user-provided builtins.

Attribute Macros

builtin