Struct jrsonnet_evaluator::EvaluationSettings

pub struct EvaluationSettings {
    pub context_initializer: TraceBox<dyn ContextInitializer>,
    pub import_resolver: TraceBox<dyn ImportResolver>,
}

Dynamically reconfigurable evaluation settings

Fields
context_initializer: TraceBox<dyn ContextInitializer>

Context initializer, which will be used for imports and everything [NoopContextInitializer] is used by default, most likely you want to have jrsonnet-stdlib

import_resolver: TraceBox<dyn ImportResolver>

Used to resolve file locations/contents