Struct jrsonnet_evaluator::State

pub struct State(/* private fields */);

Maintains stack trace and import resolution

Implementations

impl State

pub fn import_resolved_str(&self, path: SourcePath) -> Result<IStr>

Should only be called with path retrieved from [resolve_path], may panic otherwise

pub fn import_resolved_bin(&self, path: SourcePath) -> Result<IBytes>

Should only be called with path retrieved from [resolve_path], may panic otherwise

pub fn import_resolved(&self, path: SourcePath) -> Result<Val>

Should only be called with path retrieved from [resolve_path], may panic otherwise

pub fn import_from(&self, from: &SourcePath, path: &str) -> Result<Val>

Has same semantics as import 'path' called from from file

pub fn import(&self, path: impl AsRef<Path>) -> Result<Val>

pub fn create_default_context(&self, source: Source) -> Context

Creates context with all passed global variables

pub fn create_default_context_with( &self, source: Source, context_initializer: impl ContextInitializer ) -> Context

Creates context with all passed global variables, calling custom modifier

pub fn push<T>( e: CallLocation<'_>, frame_desc: impl FnOnce() -> String, f: impl FnOnce() -> Result<T> ) -> Result<T>

Executes code creating a new stack frame

pub fn push_val( &self, e: &ExprLocation, frame_desc: impl FnOnce() -> String, f: impl FnOnce() -> Result<Val> ) -> Result<Val>

Executes code creating a new stack frame

pub fn push_description<T>( frame_desc: impl FnOnce() -> String, f: impl FnOnce() -> Result<T> ) -> Result<T>

Executes code creating a new stack frame

impl State

Internals

pub fn settings(&self) -> Ref<'_, EvaluationSettings>

pub fn settings_mut(&self) -> RefMut<'_, EvaluationSettings>

pub fn add_global(&self, name: IStr, value: Thunk<Val>)

impl State

Raw methods evaluate passed values but don’t perform TLA execution

pub fn evaluate_snippet( &self, name: impl Into<IStr>, code: impl Into<IStr> ) -> Result<Val>

Parses and evaluates the given snippet

pub fn evaluate_snippet_with( &self, name: impl Into<IStr>, code: impl Into<IStr>, context_initializer: impl ContextInitializer ) -> Result<Val>

Parses and evaluates the given snippet with custom context modifier

impl State

Settings utilities

pub fn resolve_from(&self, from: &SourcePath, path: &str) -> Result<SourcePath>

pub fn resolve(&self, path: impl AsRef<Path>) -> Result<SourcePath>

pub fn import_resolver(&self) -> Ref<'_, dyn ImportResolver>

pub fn set_import_resolver(&self, resolver: impl ImportResolver)

pub fn context_initializer(&self) -> Ref<'_, dyn ContextInitializer>

pub fn set_context_initializer(&self, initializer: impl ContextInitializer)