Trait jrsonnet_evaluator::ContextInitializer

pub trait ContextInitializer: Trace {
    // Required methods
    fn populate(&self, for_file: Source, builder: &mut ContextBuilder);
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn reserve_vars(&self) -> usize { ... }
    fn initialize(&self, state: State, for_file: Source) -> Context { ... }
}

During import, this trait will be called to create initial context for file. It may initialize global variables, stdlib for example.

Required Methods

fn populate(&self, for_file: Source, builder: &mut ContextBuilder)

For composability: extend builder. May panic if this initialization is not supported, and the context may only be created via initialize.

fn as_any(&self) -> &dyn Any

Allows upcasting from abstract to concrete context initializer. jrsonnet by itself doesn’t use this method, it is allowed for it to panic.

Provided Methods

fn reserve_vars(&self) -> usize

For which size the builder should be preallocated

fn initialize(&self, state: State, for_file: Source) -> Context

Initialize default file context. Has default implementation, which calls populate. Prefer to always implement populate instead.